dataProcessScript/requirements.txt
absl-py==0.9.0
backcall==0.2.0
decorator==4.4.2
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.6
jupyter-core==4.6.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pyzmq==19.0.2
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.5
